feat: refactoring repositories, services, controllers

+++ Hospital/IntegrationAPI/DTO/ResponseToComplaintDTO.cs
@@ -7,9 +7,9 @@ namespace IntegrationAPI.DTO
 {
     public class ResponseToComplaintDTO
     {
-        public long ResponseId { get; set; }
+        public int ResponseId { get; set; }
         public DateTime Date { get; set; }
         public String Content { get; set; }
-        public long ComplaintId { get; set; }
+        public int ComplaintId { get; set; }
     }
 }
